+++ app/src/main/java/de/christian2003/chaching/plugin/presentation/view/licenses/LicensesScreen.kt
@@ -1,8 +1,8 @@
 package de.christian2003.chaching.plugin.presentation.view.licenses
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -11,22 +11,27 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SheetState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
 import de.christian2003.chaching.R
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 
 
 /**
@@ -35,13 +40,15 @@ import de.christian2003.chaching.R
  * @param viewModel     View model for the view.
  * @param onNavigateUp  Callback invoked to navigate up on the navigation stack.
  */
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun LicensesScreen(
     viewModel: LicensesViewModel,
     onNavigateUp: () -> Unit
 ) {
     Scaffold(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = MaterialTheme.colorScheme.surface),
         topBar = {
             TopAppBar(
                 title = {
@@ -83,12 +90,12 @@ fun LicensesScreen(
                 )
             }
 
-            if (viewModel.displayedLicenseName != null && viewModel.displayedLicenseText != null) {
+            if (viewModel.displayedSoftwareName != null && viewModel.displayedLicenseText != null) {
                 LicenseDialog(
-                    licenseName = viewModel.displayedLicenseName!!,
+                    softwareName = viewModel.displayedSoftwareName!!,
                     licenseText = viewModel.displayedLicenseText!!,
                     onDismiss = {
-                        viewModel.displayedLicenseName = null
+                        viewModel.displayedSoftwareName = null
                         viewModel.displayedLicenseText = null
                     }
                 )
@@ -105,7 +112,7 @@ fun LicensesScreen(
  * @param onLicenseClicked  Callback invoked once a license is clicked.
  */
 @Composable
-private fun LicensesList(
+fun LicensesList(
     licenses: List<License>,
     onLicenseClicked: (License) -> Unit
 ) {
@@ -127,11 +134,14 @@ private fun LicensesList(
  * @param onLicenseClicked  Callback invoked once the license is clicked.
  */
 @Composable
-private fun LicensesListRow(
+fun LicensesListRow(
     license: License,
     onLicenseClicked: (License) -> Unit
 ) {
-    Row(
+    Text(
+        text = license.softwareName,
+        color = MaterialTheme.colorScheme.onSurface,
+        style = MaterialTheme.typography.bodyLarge,
         modifier = Modifier
             .fillMaxWidth()
             .clickable {
@@ -140,70 +150,77 @@ private fun LicensesListRow(
             .padding(
                 vertical = dimensionResource(R.dimen.padding_vertical),
                 horizontal = dimensionResource(R.dimen.margin_horizontal)
-            ),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(
-            modifier = Modifier.weight(1f)
-        ) {
-            Text(
-                text = license.softwareName,
-                color = MaterialTheme.colorScheme.onSurface,
-                style = MaterialTheme.typography.bodyLarge
             )
-            Text(
-                text = license.licenseName,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                style = MaterialTheme.typography.bodyMedium
-            )
-        }
-        Text(
-            text = license.softwareVersion,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            style = MaterialTheme.typography.bodyMedium
-        )
-    }
+    )
 }
 
 
 /**
  * Composable displays a dialog through which a license can be displayed.
  *
- * @param licenseName       Name of the license displayed.
+ * @param softwareName      Name of the software whose license is being displayed..
  * @param licenseText       Text of the license to display.
  * @param onDismiss         Callback to close the dialog without deleting the petrol entry.
  */
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-private fun LicenseDialog(
-    licenseName: String,
+fun LicenseDialog(
+    softwareName: String,
     licenseText: String,
     onDismiss: () -> Unit
 ) {
-    val sheetState = rememberModalBottomSheetState()
+    val sheetState: SheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+    val coroutineScope: CoroutineScope = rememberCoroutineScope()
 
     ModalBottomSheet(
         onDismissRequest = onDismiss,
-        sheetState = sheetState
+        sheetState = sheetState,
+        dragHandle = null,
+        sheetGesturesEnabled = false
     ) {
         Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = dimensionResource(R.dimen.margin_horizontal))
-                .verticalScroll(rememberScrollState())
+            modifier = Modifier.fillMaxSize()
         ) {
-            Text(
-                text = licenseName,
-                color = MaterialTheme.colorScheme.onSurface,
-                style = MaterialTheme.typography.headlineSmall
+            TopAppBar(
+                title = {
+                    Text(
+                        text = softwareName,
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis
+                    )
+                },
+                navigationIcon = {
+                    IconButton(
+                        onClick = {
+                            coroutineScope.launch {
+                                sheetState.hide()
+                            }.invokeOnCompletion {
+                                onDismiss()
+                            }
+                        }
+                    ) {
+                        Icon(
+                            painter = painterResource(R.drawable.ic_cancel),
+                            contentDescription = ""
+                        )
+                    }
+                },
+                colors = TopAppBarDefaults.topAppBarColors().copy(containerColor = MaterialTheme.colorScheme.surfaceContainerLow)
             )
+
+            HorizontalDivider()
+
             Text(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = dimensionResource(R.dimen.padding_vertical)),
                 text = licenseText,
                 color = MaterialTheme.colorScheme.onSurface,
-                style = MaterialTheme.typography.bodyMedium
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .verticalScroll(rememberScrollState())
+                    .padding(
+                        start = dimensionResource(R.dimen.margin_horizontal),
+                        end = dimensionResource(R.dimen.margin_horizontal),
+                        bottom = dimensionResource(R.dimen.padding_vertical)
+                    )
             )
         }
     }