+++ app/src/main/java/de/christian2003/chaching/plugin/presentation/view/transfers/TransfersScreen.kt
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.windowInsetsBottomHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -49,7 +48,6 @@ import de.christian2003.chaching.plugin.presentation.ui.composables.TransferList
  * @param onNavigateUp      Callback invoked to navigate up on the navigation stack.
  * @param onEditTransfer    Callback invoked to navigate to the page to edit a transfer.
  */
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TransfersScreen(
     viewModel: TransfersViewModel,